+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Three buttons are included in the dialog, each of them acts as interface for the above mentioned interface methods for the sales data API
 - Simple component for Google Maps iframe embeds
   - Includes pipe for trusting external resource urls (uses `DomSanitizer`)
+- Search bar component for the frontpage header
 - Parameterizable content on the frontpage (settings can be configured using `environments`)
   - Location map (at the bottom; uses above mentioned Google Map iframe embed)
   - Banners (top and bottom of the page)
@@ -27,7 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `DataManagerComponentDirective<T>` now exposes `items$` as `Observable<any[]>` instead of `Observable<T[]>`
 - When adding images, `code` field is now optional
-- Hide frontpage heading, as the header already has a heading title
+- Hide frontpage heading
+- Replace title in frontpage header with search bar
 - `AppService` now exposes an `userName$` observable
 
 ### Fixed